[FEATURE] Vpn controller
