--- src/controllers/ipsec.rs
@@ -1,1 +1,0 @@
-
