+++ src/controllers/mod.rs
@@ -1,3 +1,4 @@
 pub mod proxy;
 pub mod dns;
 pub mod store;
+pub mod vpn;
